public pages now are in folders
