+++ src/pages/public/search/index.scss
@@ -1,0 +1,11 @@
+.search {
+  min-height: 100vh;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-image: url('../../../assets/images/20759.jpg');
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: contain;
+  flex-flow: column;
+}
